add cron job for desable boost

+++ database/migrations/2023_06_04_133900_create_annonces_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->unsignedDouble('price');
             $table->text('description');
-            $table->unsignedInteger('level')->default(1);
+            $table->unsignedBigInteger('level')->default(1);
             $table->boolean('is_blocked')->default(false);
             $table->foreignId('town_id')->constrained();
             $table->foreignId('user_id')->constrained();
